+++ Source/Core/Visualization/Mapper/OpacityMap.h
@@ -4,9 +4,7 @@
  *  @author Naohisa Sakamoto
  */
 /****************************************************************************/
-#ifndef KVS__OPACITY_MAP_H_INCLUDE
-#define KVS__OPACITY_MAP_H_INCLUDE
-
+#pragma once
 #include <list>
 #include <utility>
 #include <kvs/ValueArray>
@@ -24,40 +22,41 @@ namespace kvs
 class OpacityMap
 {
 public:
-
-    typedef kvs::ValueArray<float> Table;
-    typedef std::pair<float,float> Point;
-    typedef std::list<Point> Points;
+    using Table = kvs::ValueArray<float>;
+    using Point = std::pair<float,float>;
+    using Points = std::list<Point>;
 
 private:
-
-    size_t m_resolution; ///< table resolution
-    float m_min_value; ///< min. value
-    float m_max_value; ///< max. value
-    Points m_points; ///< control point list
-    Table m_table; ///< value table
+    size_t m_resolution = 256; ///< table resolution
+    float m_min_value = 0.0f; ///< min. value
+    float m_max_value = 0.0f; ///< max. value
+    Points m_points{}; ///< control point list
+    Table m_table{}; ///< value table
 
 public:
-
-    OpacityMap();
-    explicit OpacityMap( const size_t resolution );
-    explicit OpacityMap( const Table& table );
-    OpacityMap( const OpacityMap& other );
+    OpacityMap() = default;
+    explicit OpacityMap( const size_t resolution ): m_resolution( resolution ) {}
+    explicit OpacityMap( const Table& table ): m_resolution( table.size() ), m_table( table ) {}
+    OpacityMap( const OpacityMap& other ) { *this = other; }
     OpacityMap( const size_t resolution, const float min_value, const float max_value );
     OpacityMap( const Table& table, const float min_value, const float max_value );
-    virtual ~OpacityMap() {}
+    virtual ~OpacityMap() = default;
 
     float minValue() const { return m_min_value; }
     float maxValue() const { return m_max_value; }
     size_t resolution() const { return m_resolution; }
     const Points& points() const { return m_points; }
     const Table& table() const { return m_table; }
-
     bool hasRange() const;
+
     void setRange( const float min_value, const float max_value ) { m_min_value = min_value; m_max_value = max_value; }
     void setResolution( const size_t resolution ) { m_resolution = resolution; }
+    void setPoints( const Points& points ) { m_points = points; }
+    void setPoints( const std::list<float>& opacities );
     void addPoint( const float value, const float opacity );
     void removePoint( const float value );
+    void clearPoints() { m_points.clear(); }
+    void reversePoints() { m_points.reverse(); }
     void create();
 
     kvs::Real32 operator []( const size_t index ) const;
@@ -66,5 +65,3 @@ public:
 };
 
 } // end of namespace kvs
-
-#endif // KVS__OPACITY_MAP_H_INCLUDE